UIEH-1184 Renamed application to application-route
